refactor(#792): extract LV2 TTL parsing into dispatch_lv2_parse.rs (sweep)
dispatch.rs 651 -> 452 (leaves the debt list). The port/scale-point parsers
(parse_ports, parse_port_block, parse_port_properties, parse_scale_points,
classify, capture_after/quoted, lv2_control_value) move to dispatch_lv2_parse.rs;
the capture-grid resolver, LV2 types, and the scan entry point stay. parse_ports
and lv2_control_value are re-exported so plugin_loader::dispatch::* (lv2, project)
and super::dispatch::* (dispatch_infer, tests) resolve unchanged. plugin-loader
tests (6) pass; lv2 + project compile clean; zero warnings.
